change package name to SwiftBuilder
