--- Tests/FluentInterfaceTests/fluentinterfaceTests.swift
@@ -1,8 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by 游宗諭 on 2020/3/3.
-//
-
-import Foundation
